Fix weather effects being scaled differently in different images, and add parallax to sun effects
The reason is that the scale to fit image into screen size has been
wrongly applied to weather layers. And to fix it, we extract only the
translate part from parallax matrix to apply to weather layers. And for
foreground and background, we still use the parallax matrix.

Bug: 375272898
Bug: 375214506
Test: manually test density and thickness of raindrop across diferent
test images, and test the parallax of sun effects by swiping left and
right in homescreen while sun effects is playing
Flag: EXEMPT only change in Magic Portrait

Change-Id: I2e697dee70529291f949517172e764bb3ee090a4