Suspend [MotionValue.keepRunning] when idle #MotionMechanics
This is currently using a [SnapshotFlow] to observe the input and spring
stability. Will investigate further improvements on this, but the
current implementation works.

Also cleaned up tests a bit more, now that `waitForIdle` does
complete while autoAdvancing (it did not before, due to the endless
calls to withFrameNanos

Flag: NONE Initial commits for new library, currently unused.
Test: atest mechanics_tests
Bug: 379248269
Change-Id: I51f27332b299ec114d06ace5997960eb4f325425